slippi_labeler/requirements.txt
dearpygui
py-slippi
python-dateutil
